Add digital-thoreau-simplify watchface
Inspired by a motorola watch face and Thoreau's advice from Walden:
"Simplify, simplify." this clean and simple watch face shows the time,
date and day of week and an optionally dismissable AsteroidOS logo.
Pressing and holding on the logo causes it to disappear.  Pressing and
holding in that area again makes it reappear.

Signed-off-by: Ed Beroset <beroset@ieee.org>

fix bus