Fix pushFiles function
The previous commit added quotes around ${source} but the script
requires them to be unquoted.  This fixes issue #97 and preserves the
ability to push wallpaper from other directories, even if the directory
or file name happens to have a space in the name.  That's why the IFS=
line is there.

Signed-off-by: Ed Beroset <beroset@ieee.org>