Factor out sanitizeSourceWatchface function
This pulls some common code to check for the existence of a watchface
into a function named sanitizeSourceWatchface().  Note that the exit
from this function only terminates the subshell and sets the error code
when invoked from within $() so we must test the resulting error code in
the calling function.

Signed-off-by: Ed Beroset <beroset@ieee.org>