Fix MENUPROGRAM definition and usage
This fixes an error in which the MENUPROGRAM variable was defined
without quotes but used with them.  It should have been the other way
around instead.  Using quotes more clearly signals to the user that it
is intentional to assign a string and not the result of running the
command in this case.

Signed-off-by: Ed Beroset <beroset@ieee.org>