Add all option (#119)
* Improve style by directly testing

Rather than indirectly testing the "$?" variable, just check the result
of a command directly.

* Fix MENUPROGRAM definition and usage

This fixes an error in which the MENUPROGRAM variable was defined
without quotes but used with them.  It should have been the other way
around instead.  Using quotes more clearly signals to the user that it
is intentional to assign a string and not the result of running the
command in this case.

* Add "every" option to select all by default

This adds the "every" option so that all watchfaces are checked by
default.  This is only applicable to the "deploy" command, since no other
command can take more than one watchface.

* Remove spurious trailing spaces from some lines

* Add deployall option

This adds a non-interactive deployall function to deploy all watchfaces.

* Correct conflict resolve attempt

---------

Signed-off-by: Ed Beroset <beroset@ieee.org>
Co-authored-by: Timo Könnecke <koennecke@mosushi.de>