Refactor decimal-time to use components
This refactors the decimal-time watchface to use a few techniques.
First, it uses named inline components to simplify code and reduce
repetition.  Second, it keys general sizing from the watchface width
rather than the height to look better on a rectangular watchface that is
taller than it is wide (e.g. beluga).  Third, it eliminates the use of
an explicit z ordering in favor of simply having the z-ordering implicit
by the order of elements in the file.  Finally, it also updates the copyright
years.

Signed-off-by: Ed Beroset <beroset@ieee.org>