Add fake-components for better testing
This adds a few fake components to simulate some objects that are used
on the watch but don't typically exist on a desktop machine.

All of these work except for HrmSensor which is not yet completely
implemented, but will allow watchfaces which use it to at least run,
although with errors.

Signed-off-by: Ed Beroset <beroset@ieee.org>